feat: use loading state between sorts, move header state up
